+++ lib/instream/pool/worker.ex
@@ -53,6 +53,7 @@ defmodule Instream.Pool.Worker do
   defp execute(%Query{ type: type } = query, opts, conn) do
     case type do
       :cluster -> Query.Cluster.execute(query, opts, conn)
+      :ping    -> Query.Ping.execute(query,opts, conn)
       :read    -> Query.Read.execute(query, opts, conn)
       :write   -> Query.Write.execute(query, opts, conn)
     end