+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.9.0-dev
 
 - Enhancements
+  - Ping a connection using `Instream.Connection.ping/0`
   - Reading queries can be executed directly using `Instream.Connection.query/2`
   - Writing query can be executed directly using `Insream.Connection.write/2`